+++ Eter/Board.h
@@ -13,17 +13,14 @@ using matrix = std::vector<std::vector<std::deque<SimpleCard>>>;
 // TODO: refactoring so that I use ranges and iterators instead of basic i and j when going trough matrix
 // TODO: spune-i lui Radu sa modifice in mages sa foloseasca enum pt culoare
 // TODO: spune-i lui Radu sa faca membrii statici la mages
-// TODO: method to remove card from a position (on top)
 // TODO: refactor the code so that it uses CardColor enum instead of std::string
-// TODO: illusions
 // TODO: fix usedCard mecanic
-// TODO: fix the card placing mecanic so that it places a card on top of another only if it's bigger
 
 class Board
 {
 private:
 	matrix m_board;
-	uint8_t m_size;
+	int16_t m_size;
 public:
 	enum class State
 	{
@@ -44,12 +41,12 @@ public:
 		Right
 	};
 
-	using Position = std::tuple<uint8_t, uint8_t>;
+	using Position = std::tuple<int16_t, int16_t>;
 
 public:
 	Board(const Board& board);
 	Board();//defaul constuctor
-	Board(uint8_t size);
+	Board(int16_t size);
 	Board(Board&& board)noexcept;
 
 	Board& operator=(Board&& board)noexcept;
@@ -64,29 +61,29 @@ public:
 	void expandRow(RowExpandDirection direction);
 	void expandColumn(ColumnExpandDirection direction);
 
-	std::vector<uint8_t> searchEmptyColumns(); //returns a vector with indices to empty cols
-	std::vector<uint8_t> searchEmptyRows(); //returns a vector with indices to empty rows
-	bool canBePlaced(int8_t x, int8_t y)const;
+	std::vector<int16_t> searchEmptyColumns(); //returns a vector with indices to empty cols
+	std::vector<int16_t> searchEmptyRows(); //returns a vector with indices to empty rows
+	bool canBePlaced(int16_t x, int16_t y)const;
 
 	/*bool lineWithColor(std::string_view Color)const;
 	bool columnWithColor(std::string_view Color)const;
 	bool diagonalWithColor(std::string_view Color)const;*/
 	State checkWin();
-	int8_t sumPoints(const Color& color);
+	int16_t sumPoints(const Color& color);
 
-	uint8_t getSize()const;
-	int8_t getRowSize()const;
-	int8_t getColumnSize()const;
-	void moveSpace(uint8_t row, uint8_t column, uint8_t newRow, uint8_t newColumn);
-	void emptyRow(uint8_t row);
-	void emptyColumn(uint8_t column);
-	void removeRow(uint8_t row);
-	void removeColumn(uint8_t column);
-	void resizeBoard(uint8_t size);
+	int16_t getSize()const;
+	int16_t getRowSize()const;
+	int16_t getColumnSize()const;
+	void moveSpace(int16_t row, int16_t column, int16_t newRow, int16_t newColumn);
+	void emptyRow(int16_t row);
+	void emptyColumn(int16_t column);
+	void removeRow(int16_t row);
+	void removeColumn(int16_t column);
+	void resizeBoard(int16_t size);
 	void print()const;
 	void clear();
-	bool checkRow(uint8_t row); // checks if a row is full , for Explosions
-	bool checkColumn(uint8_t column); // checks if a column is full , for Explosions
+	bool checkRow(int16_t row); // checks if a row is full , for Explosions
+	bool checkColumn(int16_t column); // checks if a column is full , for Explosions
 	void pushCard(const SimpleCard& card, const Position& position);
 	bool canBePushed(const SimpleCard& card, const Position& position)const;
 	void popCard(const Position& position);
