+++ Eter/PowerStorm.h
@@ -16,8 +16,8 @@ public:
 	std::string getName() const;
 	std::string getDescription() const;
 
-	void playPower(Board& board, uint8_t x, uint8_t y);
-	bool checkPower(Board& board, uint8_t x, uint8_t y);
+	void playPower(Board& board, int16_t x, int16_t y);
+	bool checkPower(Board& board, int16_t x, int16_t y);
 
 };
